# BackendV2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FUTCOL API",
        "description": "API para el sistema de gestión de torneos de fútbol FUTCOL. Esta API permite gestionar torneos, equipos, jugadores, amonestaciones y estadísticas.",
        "contact": {
            "name": "FUTCOL Support Team",
            "email": "admin@futcol.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "FUTCOL API Server"
        }
    ],
    "paths": {
...
        "/api/jueces": {
            "get": {
                "tags": [
                    "Jueces"
                ],
                "summary": "Listar todos los jueces",
                "operationId": "fdddee27068db3908aacc064c43ac215",
                "responses": {
                    "200": {
                        "description": "Lista de jueces"
                    }
                }
            },
            "post": {
                "tags": [
                    "Jueces"
                ],
                "summary": "Registrar un nuevo juez",
                "operationId": "348fc8b610ff427591fdac6ff042412c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "numero_de_contacto",
                                    "correo"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos Pérez"
                                    },
                                    "numero_de_contacto": {
                                        "type": "string",
                                        "example": "3111234567"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "carlos@example.com"
                                    },
                                    "sede_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Juez creado exitosamente"
                    },
                    "422": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/jueces/{id}": {
            "get": {
                "tags": [
                    "Jueces"
                ],
                "summary": "Obtener información de un juez específico",
                "operationId": "6f5f13799c272913c00d34fbe6d1bafc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del juez"
                    },
                    "404": {
                        "description": "Juez no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jueces"
                ],
                "summary": "Actualizar información de un juez",
                "operationId": "b74dd72a45faf20224c4a320a79a5636",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Carlos Pérez"
                                    },
                                    "numero_de_contacto": {
                                        "type": "string",
                                        "example": "3111234567"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "carlos@example.com"
                                    },
                                    "sede_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Juez actualizado exitosamente"
                    },
                    "404": {
                        "description": "Juez no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jueces"
                ],
                "summary": "Eliminar un juez",
                "operationId": "4240c5b9ae2e8cb7eaaec2ea8c27367d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Juez eliminado exitosamente"
                    },
                    "404": {
                        "description": "Juez no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingresa tu token JWT en el formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}